requirements.txt
pandas==1.4.2
transformers
torchvision==0.12.0
keras==2.9.0
tensorflow==2.9.1
matplotlib==3.5.2
openpyxl== 3.0.10

